Structure: Re-arrange screen folder structure
