+++ src/screens/events/events.screen.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { View } from 'react-native';
+
+export const EventView: React.FC = () => <View />;
