+++ src/screens/events/index.tsx
@@ -1,4 +1,1 @@
-import React from 'react';
-import { View } from 'react-native';
-
-export const EventView: React.FC = () => <View />;
+export * from './events.screen';
